fixed sunday's error - was badge_set

+++ srcs/cs_u_start.c
@@ -6,7 +6,7 @@
 /*   By: jnederlo <jnederlo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/07/16 13:19:19 by jnederlo          #+#    #+#             */
-/*   Updated: 2017/07/24 15:21:42 by jnederlo         ###   ########.fr       */
+/*   Updated: 2017/07/24 15:31:38 by jnederlo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,11 +56,11 @@ int		choose_len_u(t_type *type, t_badge *badge, va_list ap)
 	}
 	else if (badge->h)
 	{
-		type->ull_int = va_arg(ap, unsigned long);
+		type->ull_int = va_arg(ap, unsigned);
 		return (num = count_digit_ulli(type));
 	}
 	else
-		type->ull_int = va_arg(ap, unsigned long);
+		type->ull_int = va_arg(ap, unsigned);
 	return (num = count_digit_ulli(type));
 }
